Move upgrade-related tests to a specific spec file
